fix clippy warnings in settings crate
