+++ services/Cargo.toml
@@ -13,18 +13,15 @@ targets = [
     "x86_64-unknown-linux-gnu",
 ]
 [dependencies]
-printnanny-asyncapi-models = { version = "0.1.15"    } # PritnNanny AsyncAPI Rust models generated by Modelina
 async-process = "1.3"
 async-trait = "0.1"
 anyhow = { version = "1", features = ["backtrace"] }
-bytes = "1"
 chrono = "0.4.22"
 config = "0.11"
 console = "0.14"
 dialoguer = "0.8"
 file-lock = "2.1.4"
 futures = "0.3"
-glob = "0.3.0"
 hex = "0.4"
 http = "0.2.5"
 jsonwebtoken = "7"
