+++ settings/Cargo.toml
@@ -10,19 +10,24 @@ description = "Utilities for interacting with PrintNanny settings"
 rust-version = "1.63"
 
 [dependencies]
+printnanny-api-client = "0.113.0"
 async-trait = "0.1"
+bytes = "1"
 printnanny-asyncapi-models = { version = "0.1.15"    } # PritnNanny AsyncAPI Rust models generated by Modelina
 serde = { version = "1", features = ["derive"] }
 clap = { version = "3", features = ["derive", "cargo", "env", "wrap_help"] }
+file-lock = "2.1.4"
 figment = { version = "0.10", features = ["env", "json", "toml", "yaml"] }
+glob = "0.3.0"
+
 log = "0.4"
 toml = "0.5"
 printnanny-dbus = { path = "../dbus", version = "^0.1" }
 serde_json = "1"
 sys-info = "0.9"
 thiserror = "1"
-
 git2 = "0.15.0"                 # Bindings to libgit2 for interoperating with git repositories.
+zip = { version = "0.6.2", default-features = false }
 
 [dev-dependencies]
 predicates = "2"
